backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
apache-beam==2.39.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==21.4.0
autopep8==1.7.0
avro-python3==1.10.2
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
bleach==5.0.0
blinker==1.5
branca==0.6.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
coloredlogs==15.0.1
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.7
coremltools==6.2
crcmod==1.7
cryptography==37.0.4
cvzone==1.5.6
cycler==0.10.0
Cython==0.29.30
dataclasses-json==0.6.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
Django==4.0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
dm-tree==0.1.7
docopt==0.6.2
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
etils==0.6.0
exceptiongroup==1.1.1
executing==0.8.2
faiss-cpu==1.7.4
fastapi==0.104.0
fastavro==1.5.1
fastjsonschema==2.15.3
filelock==3.10.7
flatbuffers==2.0.7
folium==0.14.0
fonttools==4.40.0
frozendict==2.3.7
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.56.2
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.53.0
h11==0.13.0
hdfs==2.7.0
html5lib==1.1
httplib2==0.19.1
huggingface-hub==0.17.3
humanfriendly==10.0
ibm-cos-sdk==2.11.0
ibm-cos-sdk-core==2.11.0
ibm-cos-sdk-s3transfer==2.11.0
ibm-db==3.1.3
ibm-db-sa==0.3.3
ibm-watson-machine-learning==1.0.218
idna==2.10
imageio==2.22.4
imbalanced-learn==0.9.1
importlib-resources==5.7.1
iniconfig==2.0.0
InstructorEmbedding==1.0.1
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipython-sql==0.4.1
ipywidgets==7.7.0
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.8
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.1
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
jupyterthemes==0.20.0
kaggle==1.5.12
keras==2.12.0
Keras-Preprocessing==1.1.2
keyring==23.5.1
kiwisolver==1.4.4
labelImg==1.8.6
langchain==0.0.320
langsmith==0.0.49
lesscpy==0.15.0
libs==0.0.10
lomond==0.3.3
lvis==0.5.3
lxml==4.6.4
Markdown==3.4.3
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mediapipe==0.8.11
mistune==0.8.4
ml-dtypes==0.0.4
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
networkx==2.8.8
nltk==3.8.1
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.0
object-detection @ file:///C:/Users/Gerome%20Andrew/Desktop/Projects/RealTimeObjectDetection/Tensorflow/models/research
onnx==1.13.1
onnx-coreml==1.3
onnxruntime==1.14.1
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.7.2
outcome==1.2.0
packaging==20.9
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.8.2
platformdirs==2.5.0
plotly==5.13.1
pluggy==1.0.0
ply==3.11
portalocker==2.4.0
prettytable==0.7.2
promise==2.3
prompt-toolkit==3.0.28
proto-plus==1.20.5
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.9.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.0
pydot==1.4.2
Pygments==2.11.2
pymongo==3.12.3
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypdf==3.16.4
PyPDF2==3.0.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.9.3.3
pyreadline3==3.4.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==6.1.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qiniu==7.10.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.3.0
QtPy==2.1.0
qu==1.2.5
readme-renderer==35.0
regex==2022.6.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==12.4.4
roboflow==1.1.0
rsa==4.9
sacrebleu==2.1.0
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.11.0
selenium==4.3.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.22
stack-data==0.1.4
starlette==0.27.0
statsmodels==0.13.5
streamlit==1.20.0
supervision==0.10.0
sympy==1.11.1
tabulate==0.8.9
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.17.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.12.0
tensorflow-gpu==2.9.1
tensorflow-hub==0.12.0
tensorflow-intel==2.12.0
tensorflow-io==0.26.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
tensorflow-model-optimization==0.7.2
tensorflow-object-detection-api==0.1.1
tensorflow-probability==0.19.0
tensorflow-text==2.9.0
tensorflowjs==3.18.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.9.2
tf-slim==1.1.0
tf2onnx==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2022.10.10
tinycss2==1.1.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.34.1
trio==0.21.0
trio-websocket==0.9.2
twine==4.0.0
typeguard==2.13.3
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.23.2
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==3.6.0
wordcloud==1.8.1
wsproto==1.1.0
xgboost==1.7.4
yarl==1.9.2
yfinance==0.2.17
